orchestrate: comm-aware placement (NEURA_COMM_AWARE) — volume-weighted memory proximity
Weights the placement's memory-proximity penalty by each memref's data volume
(element count) so placement minimizes sum(volume * manhattan_distance) = real
communication, instead of the uniform baseline. Gated on env NEURA_COMM_AWARE for
a controlled baseline-vs-ours A/B in one binary. Score widened to int64_t.
Real result: resnet comm 576->256 (2.25x), geomean 1.31x over real programs.

+++ include/TaskflowDialect/Orchestration/orchestration_utils.h
@@ -161,8 +161,8 @@ private:
 
   // Scores a candidate placement using proximity to dependent tasks, assigned
   // SRAMs, and context reuse cost.
-  int computeScore(TaskNode *task_node, const TaskPlacement &placement,
-                   TaskMemoryGraph &graph);
+  int64_t computeScore(TaskNode *task_node, const TaskPlacement &placement,
+                       TaskMemoryGraph &graph);
 
   int grid_rows_;
   int grid_cols_;
